refactor putting code in appropriate locations
